Creation of FireFighter and Projectiles
Created the Character Base Blueprint with an idea of wanting to change it to C++ later on. Made a child blueprint which has the firefighter. Waiting for a model to be imported. A water gun/tool has been implemented to fire a projectile. Want to changed gun to increase the power and to be held down to make it easier for players.